+++ src/apis/getNextRoutine.ts
@@ -1,7 +1,7 @@
-import { axiosCustom } from './customAxios';
+import { customAxios } from './customAxios';
 
 export async function getNextRoutine(year: number, month: number) {
-  return await axiosCustom
+  return await customAxios
     .get('tasks/list', { params: { year, month } })
     .then((res) => res.data);
 }
