+++ src/apis/tasks/deleteRoutine.ts
@@ -1,12 +1,12 @@
 // src/apis/tasks/deleteRoutine.ts
-import { axiosCustom } from '../customAxios';
+import { customAxios } from '../customAxios';
 
 export interface DeleteRoutineResponse {
   message: string;
 }
 
 export async function deleteRoutine(taskId: number) {
-  return await axiosCustom
+  return await customAxios
     .delete<DeleteRoutineResponse>(`tasks/${taskId}`)
     .then((res) => res.data);
 }
