+++ src/apis/tasks/postBannerRoutine.ts
@@ -1,10 +1,10 @@
 // src/apis/tasks/postBannerRoutine.ts
 
 import { BannerTaskInfo } from '@/types/routine';
-import { axiosCustom } from '../customAxios';
+import { customAxios } from '../customAxios';
 
 export async function postBannerRoutine(taskData: BannerTaskInfo) {
-  return await axiosCustom
+  return await customAxios
     .post<BannerTaskInfo>('tasks', taskData)
     .then((res) => res.data);
 }
