+++ src/apis/tasks/putRoutine.ts
@@ -1,9 +1,9 @@
 // src/apis/tasks/putRoutine.ts
 import { TaskInfo } from '@/types/routine';
-import { axiosCustom } from '../customAxios';
+import { customAxios } from '../customAxios';
 
 export async function putRoutine(taskData: TaskInfo) {
-  return await axiosCustom
+  return await customAxios
     .put<TaskInfo>('tasks', taskData)
     .then((res) => res.data);
 }
